Move mocks into Assets folder so they are not included in the package
