+++ Assets/SequenceFrontend/Tests/SequenceExamplesTests.asmdef
@@ -9,7 +9,9 @@
         "GUID:f7fd4ba36aabd1d499450c174865e70b",
         "GUID:f78a27d6a73d94c4baf04337e0add4dc",
         "GUID:4e0a798abbda240658187632ae443a67",
-        "GUID:b4f9c0f8f363f439b9e337f79050f189"
+        "GUID:b4f9c0f8f363f439b9e337f79050f189",
+        "GUID:e22769e76116a7e4b8fac61caa6c07d7",
+        "GUID:040286810a82b46ed9acd6d70bfbbfd4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
