+++ Assets/SequenceSDK/Authentication/Tests/MockLoginCustomEventTiming.cs
