+++ Assets/SequenceSDK/Relayer/Tests/SequenceRelayerTests.asmdef
@@ -9,7 +9,8 @@
         "GUID:6055be8ebefd69e48b49212b09b47b2f",
         "GUID:a35e3a53d4439435f8b36ed2c6158cd8",
         "GUID:94b67778c44684afdab21990eebf60aa",
-        "GUID:bcbfbf3f441b847878a6d118add93461"
+        "GUID:bcbfbf3f441b847878a6d118add93461",
+        "GUID:040286810a82b46ed9acd6d70bfbbfd4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
