+++ LocalPackages/Sequence-Unity/Sequence/SequenceFrontend/MockCurrencyConverter.cs
